ST-536: Dynamic text for when a pension has been added
