+++ app/uk/gov/hmrc/tai/service/yourTaxFreeAmount/TaxCodeChangeReasonsService.scala
@@ -19,10 +19,10 @@ package uk.gov.hmrc.tai.service.yourTaxFreeAmount
 import com.google.inject.Inject
 import play.api.i18n.Messages
 import uk.gov.hmrc.tai.model.domain.TaxCodeChange
-import uk.gov.hmrc.tai.util.yourTaxFreeAmount.{AllowancesAndDeductionPairs, EmploymentTaxCodeChangeReasons, IabdTaxCodeChangeReasons}
+import uk.gov.hmrc.tai.util.yourTaxFreeAmount.{AllowancesAndDeductionPairs, TaxCodeChangeReasons, IabdTaxCodeChangeReasons}
 
 class TaxCodeChangeReasonsService @Inject()(iabdTaxCodeChangeReasons: IabdTaxCodeChangeReasons,
-                                            employmentTaxCodeChangeReasons: EmploymentTaxCodeChangeReasons) {
+                                            employmentTaxCodeChangeReasons: TaxCodeChangeReasons) {
   def combineTaxCodeChangeReasons(iabdPairs: AllowancesAndDeductionPairs, taxCodeChange: TaxCodeChange)
                                  (implicit messages: Messages): Seq[String] = {
     val employmentReasons = employmentTaxCodeChangeReasons.reasons(taxCodeChange)
