+++ conf/messages
@@ -1394,7 +1394,8 @@ tai.updateIncome.CYPlus1.same.paragraph1=You have entered {0}
 tai.updateIncome.CYPlus1.same.paragraph2=This is the same as the amount already estimated for {0} for next year, so you do not need to update your income.
 tai.updateIncome.CYPlus1.same.cancel=Return to your Income Tax comparison for current tax year and next tax year
 
-tai.taxCodeComparison.addEmployer=You have a new employment with {0}
+tai.taxCodeComparison.add.employer=You have a new employment with {0}
+tai.taxCodeComparison.add.pension=You have a new private pension with {0}
 tai.taxCodeComparison.removeEmployer=Your employment {0} has ended
 tai.taxCodeComparison.title=We changed this because:
 tai.taxCodeComparison.iabd.updated=There has been an update to your {0}
